First kinda working version.
Lots to improve yet, but you can actually enter
guesses, have them checked and displayed, and
so on. Suspect my logic for colouring letters is
broken for words that have repeated letters, mind.
Would be nice to know the answer when you fail
too. And focus issues abound. As do aesthetic
issues.